refactor: class helpers - existeOrError
