+++ src/controllers/LocalizationController.ts
@@ -1,5 +1,5 @@
 import Localization from '@models/Localization'
-import { Request, Response } from 'express'
+import e, { Request, Response } from 'express'
 import Helpers from '@helpers/index'
 import db from '@database/connection'
 
@@ -11,9 +11,6 @@ class LocalizationController {
     } catch (msg) {
       return res.status(400).send(msg)
     }
-    // if (req.body.id) {
-    //   return res.status(401).send({ message: 'Utilize o update' })
-    // }
     const localization = new Localization()
     localization.make(req.body)
     const objectData = await localization.save()
@@ -31,24 +28,24 @@ class LocalizationController {
   }
 
   async find (req: Request, res: Response): Promise<Response> {
-    return db('localization')
-      .select('id', 'name', 'cover', 'description', 'notes')
-      .where({ id: req.params.id })
-      .first()
-      .then(localization => res.json(localization).send())
-      .catch(err => res.status(500).send(err))
+    const localization = new Localization()
+    const findLocalization = await localization.findId(req.params.id)
+    if (localization.fail.Status()) {
+      return res.status(401).send(localization.fail.Error())
+    } else {
+      return res.status(201).send(findLocalization)
+    }
   }
 
-  // async update (req: Request, res: Response): Promise<Response> {
-  //   // const localization = new Localization(req.body)
-  //   // return db('localization')
-  //   //   .update(localization)
-  //   //   .where({
-  //   //     id: req.body.id
-  //   //   })
-  //   //   .then(localization => res.status(200).send())
-  //   //   .catch(err => res.status(500).send(err))
-  // }
+  async update (req: Request, res: Response): Promise<Response> {
+    const localization = new Localization()
+    localization.make(req.body)
+    const objectData = await localization.save()
+    if (localization.fail.Status()) {
+      return res.status(401).send(localization.fail.Error())
+    }
+    return res.status(201).json(objectData)
+  }
 
   async delete (req: Request, res: Response): Promise<Response> {
     if (req.body.id == null) return res.status(401).send({ message: 'Falta o id' })
