+++ src/routes/index.ts
@@ -12,7 +12,7 @@ routes.get('/', (request, response) => {
 const localizationController = new LocalizationController()
 routes.get('/localization', localizationController.get)
 routes.post('/localization', localizationController.save)
-// routes.put('/localization', localizationController.update)
+routes.put('/localization', localizationController.update)
 routes.get('/localization/:id', localizationController.find)
 routes.delete('/localization', localizationController.delete)
